fix: move visual-identity service to protocol/identity
It is required to be called before RPC server is running on client side
