+++ node/get_status_node.go
@@ -45,7 +45,6 @@ import (
 	"github.com/status-im/status-go/services/status"
 	"github.com/status-im/status-go/services/stickers"
 	"github.com/status-im/status-go/services/subscriptions"
-	visualIdentity "github.com/status-im/status-go/services/visual-identity"
 	"github.com/status-im/status-go/services/wakuext"
 	"github.com/status-im/status-go/services/wakuv2ext"
 	"github.com/status-im/status-go/services/wallet"
@@ -118,7 +117,6 @@ type StatusNode struct {
 	gifSrvc                *gif.Service
 	stickersSrvc           *stickers.Service
 	chatSrvc               *chat.Service
-	visualIdentitySrvc     *visualIdentity.Service
 }
 
 // New makes new instance of StatusNode.
@@ -429,7 +427,6 @@ func (n *StatusNode) stop() error {
 	n.wakuV2ExtSrvc = nil
 	n.ensSrvc = nil
 	n.stickersSrvc = nil
-	n.visualIdentitySrvc = nil
 	n.publicMethods = make(map[string]bool)
 
 	return nil
